Move dialogs to new directory
