+++ frontend/src/components/course/ActiveCourse.js
@@ -8,9 +8,9 @@ import { Edit as EditIcon } from '@material-ui/icons'
 import { DELETE_CONCEPT } from '../../graphql/Mutation'
 import { COURSE_BY_ID } from '../../graphql/Query'
 import ActiveConcept from '../concept/ActiveConcept'
-import useCreateConceptDialog from './useCreateConceptDialog'
-import useEditConceptDialog from './useEditConceptDialog'
-import useEditCourseDialog from './useEditCourseDialog'
+import useCreateConceptDialog from '../../dialogs/concept/useCreateConceptDialog'
+import useEditConceptDialog from '../../dialogs/concept/useEditConceptDialog'
+import useEditCourseDialog from '../../dialogs/course/useEditCourseDialog'
 import { useLoginStateValue } from '../../store'
 import { useInfoBox } from '../common/InfoBox'
