+++ frontend/src/dialogs/concept/useEditConceptDialog.js
@@ -3,7 +3,7 @@ import { useMutation } from 'react-apollo-hooks'
 
 import { UPDATE_CONCEPT } from '../../graphql/Mutation'
 import { updateConceptUpdate } from '../../apollo/update'
-import ConceptEditingDialog from '../concept/ConceptEditingDialog'
+import ConceptEditingDialog from './ConceptEditingDialog'
 
 const useEditConceptDialog = (activeCourse, workspaceId) => {
   const [conceptEditState, setConceptEditState] = useState({
