+++ frontend/src/dialogs/course/CourseCreationDialog.js
