+++ frontend/src/dialogs/course/CourseEditingDialog.js
