+++ frontend/src/dialogs/course/useEditCourseDialog.js
