+++ frontend/src/dialogs/project/ProjectEditingDialog.js
