+++ frontend/src/dialogs/project/ProjectSharingDialog.js
