+++ frontend/src/dialogs/project/TemplateCreationDialog.js
