+++ frontend/src/dialogs/workspace/WorkspaceCreationDialog.js
