+++ frontend/src/dialogs/workspace/WorkspaceEditingDialog.js
