+++ frontend/src/dialogs/workspace/WorkspaceSharingDialog.js
