+++ frontend/src/dialogs/workspace/useEditWorkspaceDialog.js
@@ -3,7 +3,7 @@ import { useMutation } from 'react-apollo-hooks'
 
 import { UPDATE_WORKSPACE } from '../../graphql/Mutation'
 import { WORKSPACES_FOR_USER, WORKSPACE_BY_ID } from '../../graphql/Query'
-import WorkspaceEditingDialog from '../workspace/WorkspaceEditingDialog'
+import WorkspaceEditingDialog from './WorkspaceEditingDialog'
 
 const useEditWorkspaceDialog = (workspaceId) => {
   const [workspaceEditState, setWorkspaceEditState] = useState({
