+++ frontend/src/dialogs/workspace/useShareWorkspaceDialog.js
